Remove CSS filter
I'm not really sure the purpose of this. When I use this in my project I end
up with absolute paths to my images which are certainly not resolvable by the
browser. I am guessing this means I am not passing the paths correct to the
Pipeline.

But even if I was I'm not sure changing the url is the right thing to do. This
assumes that everything in the asset pipeline is servable on the site. The
asset pipeline could be pulling from some none-standard location.

I think the correct method would be to use PHP as part of the pipeline to return
a correct pathname if that is what you desire. But if you just use CSS's url
then I think the correct behavior is to leave it untouched.